requirements.txt
twisted>=22.2.0
pathlib>=1.0.1
requests>=2.31.0
concurrent-log-handler
insanely-fast-whisper
transformers
optimum
accelerate
python-telegram-bot
# openai-whisper
# torch
# torchvision
# torchaudio
# git+https://github.com/huggingface/transformers
# accelerate
# optimum
# click
